+++ src/lib.rs
@@ -1815,7 +1815,7 @@ pub mod prelude {
         value::ReprValue as _,
     };
 }
-mod r_array;
+pub mod r_array;
 mod r_bignum;
 mod r_complex;
 mod r_file;